+++ pytorch_segmentation_models_trainer/custom_callbacks/__init__.py
@@ -1,4 +1,5 @@
 from .image_callbacks import EnhancedImageSegmentationResultCallback
+from .kl_annealing_callback import KLAnnealingCallback
 from .metrics_callbacks import ConfusionMatrixCallback
 from .training_callbacks import EMACallback
 
@@ -6,4 +7,5 @@ __all__ = [
     "ConfusionMatrixCallback",
     "EnhancedImageSegmentationResultCallback",
     "EMACallback",
+    "KLAnnealingCallback",
 ]